(fix) update entity

+++ src/main/java/com/stormx/hicoder/services/implement/PostServiceImpl.java
@@ -1,6 +1,7 @@
 package com.stormx.hicoder.services.implement;
 
 import com.stormx.hicoder.controllers.requests.NewPostRequest;
+import com.stormx.hicoder.dto.PostDTO;
 import com.stormx.hicoder.entities.Post;
 import com.stormx.hicoder.entities.User;
 import com.stormx.hicoder.exceptions.BadRequestException;
@@ -19,8 +20,11 @@ public class PostServiceImpl implements PostService {
     private final UserRepository userRepository;
 
     @Override
-    public List<Post> getAllPostsOfUser(User user) {
-        return postRepository.findAllByUser(user);
+    public List<PostDTO> getAllPostsOfUser(User user) {
+        List<Post> userPosts = postRepository.findAllByAuthor(user);
+        return userPosts.stream()
+                .map(PostDTO::new)
+                .toList();
     }
 
     @Override
@@ -28,31 +32,48 @@ public class PostServiceImpl implements PostService {
         return postRepository.findById(postId).orElseThrow(() -> new BadRequestException("Post not found id: " + postId));
     }
 
-    @Override
-    public Post createPost(NewPostRequest post, User user) {
-        Post newPost = Post.builder()
-                .title(post.getTitle())
-                .content(post.getContent())
-                .author(user)
-                .build();
 
-        return postRepository.save(newPost);
+    @Override
+    public PostDTO getUserPostById(String postId, User currentUser) {
+        Post userPost = getPostById(postId);
+        if (!userPost.isPostedBy(currentUser)) {
+            throw new BadRequestException("User doesn't have post: " + postId);
+        }
+        return new PostDTO(userPost);
+    }
+
+    @Override
+    public PostDTO createPost(NewPostRequest newPostRequest, User user) {
+        Post newPost = new Post(newPostRequest);
+        newPost.setAuthor(user);
+        user.addPost(newPost);
+        userRepository.save(user);
+        postRepository.save(newPost);
+        return new PostDTO(newPost);
     }
 
 
     @Override
-    public Post updatePost(String postId, NewPostRequest postDetails) {
-        Post post = getPostById(postId);
-        post.setTitle(postDetails.getTitle());
-        post.setContent(postDetails.getContent());
-        return postRepository.save(post);
+    public PostDTO updatePost(String postId, NewPostRequest postDetails, User currentUser) {
+        Post userPosts = getPostById(postId);
+        if (!userPosts.isPostedBy(currentUser)) {
+            throw new BadRequestException("User doesn't have post: " + postId);
+        }
+        userPosts.setTitle(postDetails.getTitle());
+        userPosts.setContent(postDetails.getContent());
+        postRepository.save(userPosts);
+        return new PostDTO(userPosts);
     }
 
     @Override
-    public boolean deletePost(String postId) {
+    public void deletePost(String postId, User currentUser) {
         Post post = getPostById(postId);
+        if (!post.isPostedBy(currentUser)) {
+            throw new BadRequestException("User doesn't have post: " + postId);
+        }
+        currentUser.removePost(post);
+        userRepository.save(currentUser);
         postRepository.delete(post);
-        return true;
     }
 
 }
